event kinds as int, not uint8.

+++ event/event.go
@@ -14,11 +14,11 @@ import (
 )
 
 const (
-	KindSetMetadata            uint8 = 0
-	KindTextNote               uint8 = 1
-	KindRecommendServer        uint8 = 2
-	KindContactList            uint8 = 3
-	KindEncryptedDirectMessage uint8 = 4
+	KindSetMetadata            int = 0
+	KindTextNote               int = 1
+	KindRecommendServer        int = 2
+	KindContactList            int = 3
+	KindEncryptedDirectMessage int = 4
 )
 
 type Event struct {
